✨ (core): 异常处理逻辑完善

+++ packages/core/src/core.ts
@@ -2,12 +2,13 @@ import { Knex } from 'knex'
 import { SimpleGit } from 'simple-git'
 import path from 'node:path'
 import fs from 'node:fs'
-import { pwd, cd, exec, set } from 'shelljs'
+import { pwd, exec } from 'shelljs'
 import { initDB, initDBInstance } from './db'
 import { cloneRemoteRepoToLocal, getGitWorkDIR, initGitInstance } from './git'
 import { genRandomLowercaseString } from './utils'
 import defaultDockerComposeConfig from './default-docker-compose.json'
 import { stringify } from 'yaml'
+import { safeCD } from './shellWrapper'
 
 class Core {
   git: SimpleGit
@@ -23,7 +24,13 @@ class Core {
 
   async clone(config: { project: Project.InputParam; user: User.InputParam }) {
     const { project, user } = config
-    const dirName = await cloneRemoteRepoToLocal(this.git, project, user)
+    let dirName
+    try {
+      dirName = await cloneRemoteRepoToLocal(this.git, project, user)
+    } catch (err) {
+      console.error('Project clone error')
+      return
+    }
     const ids = await this.Projects().insert({
       ...project,
       dirName
@@ -70,23 +77,32 @@ class Core {
     console.log('project info', project)
     console.log(`current work dir: ${pwd().toString()}`)
     const workDir = path.join(getGitWorkDIR(), project.dirName, project.subWorkDir || '')
-    cd(workDir)
+    try {
+      safeCD(workDir)
+    } catch (err) {
+      throw new Error('项目的 subWorkDir 有误')
+    }
     this.git.cwd(workDir)
     console.log(`current work dir: ${pwd().toString()}`)
     const rs = genRandomLowercaseString()
 
     switch (project.runMode) {
       case 'dockerfile': {
+        if (!fs.existsSync('Dockerfile')) throw new Error('Dockerfile 文件不存在')
         const serviceName = `${project.appName}-${rs}`
         console.log('serviceName', serviceName)
-        exec(`export COMPOSE_PROJECT_NAME=${serviceName}`)
         const domainName = await this.genDomainName(project)
         console.log('domainName', domainName)
         const config = await this.genDockerComposeConfig(serviceName, domainName)
         fs.writeFileSync('./docker-compose.yaml', stringify(config))
         console.log('** build docker image start **')
-        exec(`docker build -t ${serviceName} .`)
-        exec(`docker-compose --project-name ${domainName} up -d`)
+        if (exec(`docker build -t ${serviceName} .`, { silent: true }).code !== 0) {
+          throw new Error('docker build fail')
+        }
+        console.log('** build docker image success **')
+        if (exec(`docker-compose --project-name ${domainName} up -d`).code !== 0) {
+          throw new Error('Failed to start app on dockerfile mode')
+        }
         break
       }
       default:
